+++ plugins/postcss-double-position-gradients/src/has-supports-at-rule-ancestor.ts
@@ -1,6 +1,7 @@
 import { includesGradientsFunction } from './is-gradient';
+import type { AtRule, Node } from 'postcss';
 
-export function hasSupportsAtRuleAncestor(node) {
+export function hasSupportsAtRuleAncestor(node: Node): boolean {
 	let parent = node.parent;
 	while (parent) {
 		if (parent.type !== 'atrule') {
@@ -8,7 +9,7 @@ export function hasSupportsAtRuleAncestor(node) {
 			continue;
 		}
 
-		if (parent.name === 'supports' && includesGradientsFunction(parent.params)) {
+		if ((parent as AtRule).name === 'supports' && includesGradientsFunction((parent as AtRule).params)) {
 			return true;
 		}